flask_mysql_api/requirements.txt
flask
flask-mysql
flask-cors